requirements.txt
langchain==0.1.0
langgraph==0.2.0
langchain-openai==0.1.0
langchain-community==0.0.32
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
tavily-python==0.3.3
yfinance==0.2.28
reportlab==4.0.7
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
asyncio
aiohttp==3.9.1
python-dateutil==2.8.2
pydantic==2.5.2
typing-extensions==4.8.0